Consolidate all code intel testing code.
