+++ web/src/regression/codeintel.test.ts
@@ -2,31 +2,25 @@
  * @jest-environment node
  */
 
-import * as GQL from '../../../shared/src/graphql/schema'
-import { Driver } from '../../../shared/src/e2e/driver'
-import { enableLSIF, uploadAndEnsure, disableLSIF, clearUploads } from './util/codeintel'
-import { ensureLoggedInOrCreateTestUser } from './util/helpers'
-import { ensureTestExternalService, getUser, setUserSiteAdmin } from './util/api'
-import { getConfig } from '../../../shared/src/e2e/config'
+import * as path from 'path'
+import { range } from 'lodash'
+import { ElementHandle } from 'puppeteer'
+import { map } from 'rxjs/operators'
+import * as child_process from 'mz/child_process'
+import { applyEdits } from '@sqs/jsonc-parser'
+import { setProperty } from '@sqs/jsonc-parser/lib/edit'
 import { getTestTools } from './util/init'
 import { GraphQLClient } from './util/GraphQLClient'
-import { testCodeNavigation } from './util/codenav'
 import { TestResourceManager } from './util/TestResourceManager'
+import { ensureTestExternalService, getUser, setUserSiteAdmin } from './util/api'
+import { ensureLoggedInOrCreateTestUser, getGlobalSettings } from './util/helpers'
+import * as GQL from '../../../shared/src/graphql/schema'
+import { Driver } from '../../../shared/src/e2e/driver'
+import { Config, getConfig } from '../../../shared/src/e2e/config'
+import { dataOrThrowErrors, gql } from '../../../shared/src/graphql/graphql'
+import { overwriteSettings } from '../../../shared/src/settings/edit'
 
 describe('Code intelligence regression test suite', () => {
-    const prometheusCommonHeadCommit = 'b5fe7d854c42dc7842e48d1ca58f60feae09d77b' // HEAD
-    const prometheusCommonLSIFCommit = '287d3e634a1e550c9e463dd7e5a75a422c614505' // 2 behind HEAD
-    const prometheusCommonFallbackCommit = 'e8215224146358493faab0295ce364cd386223b9' // 2 behind LSIF
-    const prometheusClientHeadCommit = '333f01cef0d61f9ef05ada3d94e00e69c8d5cdda'
-    const prometheusRedefinitionsHeadCommit = 'c68f0e063cf8a98e7ce3428cfd50588746010f1f'
-
-    const testRepoSlugs = [
-        'sourcegraph/sourcegraph',
-        'sourcegraph-testing/prometheus-common',
-        'sourcegraph-testing/prometheus-client-golang',
-        'sourcegraph-testing/prometheus-redefinitions',
-    ]
-
     const testUsername = 'test-sg-codeintel'
     const config = getConfig(
         'gitHubToken',
@@ -45,6 +39,64 @@ describe('Code intelligence regression test suite', () => {
         kind: GQL.ExternalServiceKind.GITHUB,
         uniqueDisplayName: '[TEST] GitHub (codeintel.test.ts)',
     }
+
+    const testRepoSlugs = [
+        'sourcegraph/sourcegraph',
+        'sourcegraph-testing/prometheus-common',
+        'sourcegraph-testing/prometheus-client-golang',
+        'sourcegraph-testing/prometheus-redefinitions',
+    ]
+
+    const prometheusCommonHeadCommit = 'b5fe7d854c42dc7842e48d1ca58f60feae09d77b' // HEAD
+    const prometheusCommonLSIFCommit = '287d3e634a1e550c9e463dd7e5a75a422c614505' // 2 behind HEAD
+    const prometheusCommonFallbackCommit = 'e8215224146358493faab0295ce364cd386223b9' // 2 behind LSIF
+    const prometheusClientHeadCommit = '333f01cef0d61f9ef05ada3d94e00e69c8d5cdda'
+    const prometheusRedefinitionsHeadCommit = 'c68f0e063cf8a98e7ce3428cfd50588746010f1f'
+
+    const prometheusCommonSamplePairLocations = [
+        { path: 'model/value.go', line: 31, character: 19 },
+        { path: 'model/value.go', line: 78, character: 6 },
+        { path: 'model/value.go', line: 84, character: 9 },
+        { path: 'model/value.go', line: 97, character: 10 },
+        { path: 'model/value.go', line: 104, character: 10 },
+        { path: 'model/value.go', line: 108, character: 9 },
+        { path: 'model/value.go', line: 137, character: 43 },
+        { path: 'model/value.go', line: 147, character: 10 },
+        { path: 'model/value.go', line: 163, character: 10 },
+        { path: 'model/value.go', line: 225, character: 11 },
+        { path: 'model/value_test.go', line: 133, character: 9 },
+        { path: 'model/value_test.go', line: 137, character: 11 },
+        { path: 'model/value_test.go', line: 156, character: 10 },
+    ]
+
+    const prometheusClientSamplePairLocations = [
+        { path: 'api/prometheus/v1/api.go', line: 41, character: 15 },
+        { path: 'api/prometheus/v1/api.go', line: 70, character: 17 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1119, character: 18 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1123, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1127, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1131, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1135, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1139, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1143, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1147, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1151, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1155, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1159, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1163, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1167, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1171, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1175, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1179, character: 20 },
+        { path: 'api/prometheus/v1/api_test.go', line: 1197, character: 17 },
+        { path: 'api/prometheus/v1/api_bench_test.go', line: 34, character: 26 },
+    ]
+
+    const prometheusRedefinitionSamplePairLocations = [
+        { path: 'sample.go', line: 7, character: 6 },
+        { path: 'sample.go', line: 12, character: 10 },
+        { path: 'sample.go', line: 16, character: 10 },
+    ]
 
     let driver: Driver
     let gqlClient: GraphQLClient
@@ -84,8 +136,7 @@ describe('Code intelligence regression test suite', () => {
             throw new Error(`test user ${testUsername} does not exist`)
         }
         await setUserSiteAdmin(gqlClient, user.id, true)
-    }, 30 * 1000)
-
+    }, 30 * 1000) // sourcegraph/sourcegraph takes a while to clone
     afterAll(async () => {
         if (!config.noCleanup) {
             await outerResourceManager.destroyAll()
@@ -98,7 +149,7 @@ describe('Code intelligence regression test suite', () => {
     describe('Basic code intelligence regression test suite', () => {
         const innerResourceManager = new TestResourceManager()
         beforeAll(async () => {
-            innerResourceManager.add('Global setting', 'codeIntel.lsif', await disableLSIF(gqlClient))
+            innerResourceManager.add('Global setting', 'codeIntel.lsif', await setGlobalLSIFSetting(gqlClient, false))
         })
         afterAll(async () => {
             if (!config.noCleanup) {
@@ -106,52 +157,25 @@ describe('Code intelligence regression test suite', () => {
             }
         })
 
-        test('Definitions, references, and hovers', async () => {
-            const makeTestCase = (
-                repository: string,
-                sourceCommit: string,
-                /** The commit to use for the definition (changes with nearest commit). */
-                commonCommit: string,
-                path: string,
-                line: number
-            ) => ({
-                repoRev: `github.com/sourcegraph-testing/${repository}@${sourceCommit}`,
-                files: [
+        test('Definitions, references, and hovers', () =>
+            testCodeNavigation(driver, config, {
+                page: `/github.com/sourcegraph-testing/prometheus-client-golang@${prometheusClientHeadCommit}/-/blob/api/prometheus/v1/api.go`,
+                line: 41,
+                token: 'SamplePair',
+                precise: false,
+                expectedHoverContains: 'SamplePair pairs a SampleValue with a Timestamp.',
+                expectedDefinition: [
                     {
-                        path,
-                        locations: [
-                            {
-                                line,
-                                token: 'SamplePair',
-                                precise: false,
-                                expectedHoverContains: 'SamplePair pairs a SampleValue with a Timestamp.',
-                                expectedDefinition: [
-                                    {
-                                        url: `/github.com/sourcegraph-testing/prometheus-common@${commonCommit}/-/blob/model/value.go#L78:1`,
-                                        precise: false,
-                                    },
-                                    {
-                                        url: `/github.com/sourcegraph-testing/prometheus-redefinitions@${prometheusRedefinitionsHeadCommit}/-/blob/sample.go#L7:1`,
-                                        precise: false,
-                                    },
-                                ],
-                                expectedReferences: [],
-                            },
-                        ],
+                        url: `/github.com/sourcegraph-testing/prometheus-common@${prometheusCommonHeadCommit}/-/blob/model/value.go#L78:1`,
+                        precise: false,
+                    },
+                    {
+                        url: `/github.com/sourcegraph-testing/prometheus-redefinitions@${prometheusRedefinitionsHeadCommit}/-/blob/sample.go#L7:1`,
+                        precise: false,
                     },
                 ],
-            })
-
-            await testCodeNavigation(driver, config, [
-                makeTestCase(
-                    'prometheus-client-golang',
-                    prometheusClientHeadCommit,
-                    prometheusCommonHeadCommit,
-                    '/api/prometheus/v1/api.go',
-                    41
-                ),
-            ])
-        })
+                expectedReferences: [],
+            }))
 
         test(
             'File sidebar, multiple levels of directories',
@@ -237,7 +261,7 @@ describe('Code intelligence regression test suite', () => {
             }
 
             await clearUploads(gqlClient, 'github.com/sourcegraph-testing/prometheus-redefinitions')
-            innerResourceManager.add('Global setting', 'codeIntel.lsif', await enableLSIF(gqlClient))
+            innerResourceManager.add('Global setting', 'codeIntel.lsif', await setGlobalLSIFSetting(gqlClient, true))
         }, 30 * 1000)
         afterAll(async () => {
             if (!config.noCleanup) {
@@ -245,163 +269,547 @@ describe('Code intelligence regression test suite', () => {
             }
         })
 
-        test(
-            'Cross-repository definitions, references, and hovers',
-            async () => {
-                interface Location {
-                    path: string
-                    position: { line: number; character: number }
-                }
+        /**
+         * Construct a code navigation test case that ensures that there are precise results from
+         * the prometheus-common and prometheus-client-golang repositories and imprecise results from
+         * the prometheus-redefinitions library.
+         *
+         * All of these tests deal with the same SamplePair struct.
+         */
+        const makeTestCase = (page: string, line: number): CodeNavigationTestCase => {
+            const prometheusCommonPrefix = `/github.com/sourcegraph-testing/prometheus-common@${prometheusCommonLSIFCommit}/-/blob/`
+            const prometheusCommonLocations = prometheusCommonSamplePairLocations.map(({ path, line, character }) => ({
+                url: `${prometheusCommonPrefix}${path}#L${line}:${character}`,
+                precise: true,
+            }))
 
-                const commonLocations = [
-                    { path: 'model/value.go', position: { line: 31, character: 19 } },
-                    { path: 'model/value.go', position: { line: 78, character: 6 } },
-                    { path: 'model/value.go', position: { line: 84, character: 9 } },
-                    { path: 'model/value.go', position: { line: 97, character: 10 } },
-                    { path: 'model/value.go', position: { line: 104, character: 10 } },
-                    { path: 'model/value.go', position: { line: 108, character: 9 } },
-                    { path: 'model/value.go', position: { line: 137, character: 43 } },
-                    { path: 'model/value.go', position: { line: 147, character: 10 } },
-                    { path: 'model/value.go', position: { line: 163, character: 10 } },
-                    { path: 'model/value.go', position: { line: 225, character: 11 } },
-                    { path: 'model/value_test.go', position: { line: 133, character: 9 } },
-                    { path: 'model/value_test.go', position: { line: 137, character: 11 } },
-                    { path: 'model/value_test.go', position: { line: 156, character: 10 } },
+            const prometheusClientPrefix = `/github.com/sourcegraph-testing/prometheus-client-golang@${prometheusClientHeadCommit}/-/blob/`
+            const prometheusClientLocations = prometheusClientSamplePairLocations.map(({ path, line, character }) => ({
+                url: `${prometheusClientPrefix}${path}#L${line}:${character}`,
+                precise: true,
+            }))
 
-                    // These (should) also come back from the backend, but the UI merges
-                    // these into one of the results above.
-                    // { path: 'model/value.go', position: { line: 104, character: 31 } },
-                    // { path: 'model/value.go', position: { line: 150, character: 10 } },
-                    // { path: 'model/value.go', position: { line: 166, character: 10 } },
-                ]
-
-                const clientLocations = [
-                    { path: 'api/prometheus/v1/api.go', position: { line: 41, character: 15 } },
-                    { path: 'api/prometheus/v1/api.go', position: { line: 70, character: 17 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1119, character: 18 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1123, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1127, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1131, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1135, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1139, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1143, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1147, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1151, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1155, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1159, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1163, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1167, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1171, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1175, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1179, character: 20 } },
-                    { path: 'api/prometheus/v1/api_test.go', position: { line: 1197, character: 17 } },
-                    { path: 'api/prometheus/v1/api_bench_test.go', position: { line: 34, character: 26 } },
-
-                    // These (should) also come back from the backend, but the UI merges
-                    // these into one of the results above.
-                    // { path: 'api/prometheus/v1/api_bench_test.go', position: { line: 37, character: 24 } },
-                ]
-
-                const redefinitionLocations = [
-                    { path: 'sample.go', position: { line: 7, character: 6 } },
-                    { path: 'sample.go', position: { line: 12, character: 10 } },
-                    { path: 'sample.go', position: { line: 16, character: 10 } },
-                ]
-
-                const makePath = (
-                    repository: string,
-                    commit: string,
-                    { path, position: { line, character } }: Location
-                ) => `/github.com/sourcegraph-testing/${repository}@${commit}/-/blob/${path}#L${line}:${character}`
-
-                const makeTestCase = (
-                    repository: string,
-                    sourceCommit: string,
-                    /** The commit to use for the definition (changes with nearest commit). */
-                    commonCommit: string,
-                    path: string,
-                    line: number
-                ) => ({
-                    repoRev: `github.com/sourcegraph-testing/${repository}@${sourceCommit}`,
-                    files: [
-                        {
-                            path,
-                            locations: [
-                                {
-                                    line,
-                                    token: 'SamplePair',
-                                    precise: true,
-                                    expectedHoverContains: 'SamplePair pairs a SampleValue with a Timestamp.',
-                                    expectedDefinition: {
-                                        url: `/github.com/sourcegraph-testing/prometheus-common@${commonCommit}/-/blob/model/value.go#L78:6`,
-                                        precise: true,
-                                    },
-                                    expectedReferences: commonLocations
-                                        .map(r => ({
-                                            url: makePath('prometheus-common', commonCommit, r),
-                                            precise: true,
-                                        }))
-                                        .concat(
-                                            clientLocations.map(r => ({
-                                                url: makePath(
-                                                    'prometheus-client-golang',
-                                                    prometheusClientHeadCommit,
-                                                    r
-                                                ),
-                                                precise: true,
-                                            }))
-                                        )
-                                        .concat(
-                                            redefinitionLocations.map(r => ({
-                                                url: makePath(
-                                                    'prometheus-redefinitions',
-                                                    prometheusRedefinitionsHeadCommit,
-                                                    r
-                                                ),
-                                                precise: false,
-                                            }))
-                                        ),
-                                },
-                            ],
-                        },
-                    ],
+            const prometheusRedefinitionPrefix = `/github.com/sourcegraph-testing/prometheus-redefinitions@${prometheusRedefinitionsHeadCommit}/-/blob/`
+            const prometheusRedefinitionLocations = prometheusRedefinitionSamplePairLocations.map(
+                ({ path, line, character }) => ({
+                    url: `${prometheusRedefinitionPrefix}${path}#L${line}:${character}`,
+                    precise: false,
                 })
+            )
 
-                await testCodeNavigation(driver, config, [
-                    makeTestCase(
-                        'prometheus-common',
-                        prometheusCommonLSIFCommit,
-                        prometheusCommonLSIFCommit,
-                        '/model/value.go',
-                        31
-                    ),
-                    makeTestCase(
-                        'prometheus-common',
-                        prometheusCommonLSIFCommit,
-                        prometheusCommonLSIFCommit,
-                        '/model/value_test.go',
-                        133
-                    ),
-                    makeTestCase(
-                        'prometheus-client-golang',
-                        prometheusClientHeadCommit,
-                        prometheusCommonLSIFCommit,
-                        '/api/prometheus/v1/api.go',
-                        41
-                    ),
-                    // makeTestCase('prometheus-client-golang', clientCommit, commonCommit, '/api/prometheus/v1/api_test.go', 1119),
-                    // makeTestCase('prometheus-client-golang', clientCommit, commonCommit, '/api/prometheus/v1/api_bench_test.go', 34),
-                    makeTestCase(
-                        'prometheus-common',
-                        prometheusCommonFallbackCommit,
-                        prometheusCommonLSIFCommit,
-                        '/model/value.go',
-                        31
-                    ),
-                    // makeTestCase('prometheus-common', olderCommonCommit, olderCommonCommit, '/model/value.go', 31),
-                ])
-            },
-            60 * 1000
-        )
+            return {
+                page,
+                line,
+                token: 'SamplePair',
+                precise: true,
+                expectedHoverContains: 'SamplePair pairs a SampleValue with a Timestamp.',
+                expectedDefinition: {
+                    url: `/github.com/sourcegraph-testing/prometheus-common@${prometheusCommonLSIFCommit}/-/blob/model/value.go#L78:6`,
+                    precise: true,
+                },
+                expectedReferences: prometheusCommonLocations
+                    .concat(prometheusClientLocations)
+                    .concat(prometheusRedefinitionLocations),
+            }
+        }
+
+        test('Cross-repository definitions, references, and hovers (from definition)', async () => {
+            await testCodeNavigation(
+                driver,
+                config,
+                makeTestCase(
+                    `/github.com/sourcegraph-testing/prometheus-common@${prometheusCommonLSIFCommit}/-/blob/model/value.go`,
+                    31
+                )
+            )
+        })
+
+        test('Cross-repository definitions, references, and hovers (from reference)', async () => {
+            await testCodeNavigation(
+                driver,
+                config,
+                makeTestCase(
+                    `/github.com/sourcegraph-testing/prometheus-common@${prometheusCommonLSIFCommit}/-/blob/model/value_test.go`,
+                    133
+                )
+            )
+        })
+
+        test('Cross-repository definitions, references, and hovers (from remote reference)', async () => {
+            await testCodeNavigation(
+                driver,
+                config,
+                makeTestCase(
+                    `/github.com/sourcegraph-testing/prometheus-client-golang@${prometheusClientHeadCommit}/-/blob/api/prometheus/v1/api.go`,
+                    41
+                )
+            )
+        })
+
+        test('Cross-repository definitions, references, and hovers (from old commit)', async () => {
+            await testCodeNavigation(
+                driver,
+                config,
+                makeTestCase(
+                    `/github.com/sourcegraph-testing/prometheus-common@${prometheusCommonFallbackCommit}/-/blob/model/value.go`,
+                    31
+                )
+            )
+        })
     })
 })
+
+//
+// Code navigation utilities
+
+export interface CodeNavigationTestCase {
+    /**
+     * The source page.
+     */
+    page: string
+
+    /**
+     * The source line.
+     */
+    line: number
+
+    /**
+     * The token to click. Should be unambiguous within this line for the test to succeed.
+     */
+    token: string
+
+    /**
+     * Whether or not definition/hover results are precise
+     */
+    precise: boolean
+
+    /**
+     * A substring of the expected hover text
+     */
+    expectedHoverContains: string
+
+    /**
+     * A locations (if unambiguous), or a subset of locations that must occur within the definitions panel.
+     */
+    expectedDefinition: TestLocation | TestLocation[]
+
+    /**
+     * A subset of locations that must occur within the references panel.
+     */
+    expectedReferences?: TestLocation[]
+}
+
+export interface TestLocation {
+    url: string
+
+    /**
+     * Whether or not this location should be accompanied by a UI badge indicating imprecise code intel. Precise = no badge.
+     */
+    precise: boolean
+}
+
+/**
+ * Navigate to the given page and test the definitions, references, and hovers of the token
+ * on the given line. Will ensure both hover and clicking the token produces the hover overlay.
+ * Will check the precision indicator of the hoverlay and each file match in the definition
+ * and reference panels. Will compare hover text. Will compare location of each file match or
+ * the target of the page navigated to on jump-to-definition (in the case of a single definition).
+ */
+export async function testCodeNavigation(
+    driver: Driver,
+    config: Pick<Config, 'sourcegraphBaseUrl'>,
+    {
+        page,
+        line,
+        token,
+        precise,
+        expectedHoverContains,
+        expectedDefinition,
+        expectedReferences,
+    }: CodeNavigationTestCase
+): Promise<void> {
+    await driver.page.goto(config.sourcegraphBaseUrl + page)
+    await driver.page.waitForSelector('.e2e-blob')
+    const tokenEl = await findTokenElement(driver, line, token)
+
+    // Check hover
+    await tokenEl.hover()
+    await waitForHover(driver, expectedHoverContains, precise)
+
+    // Check click
+    await clickOnEmptyPartOfCodeView(driver)
+    await tokenEl.click()
+    await waitForHover(driver, expectedHoverContains)
+
+    // Find-references
+    if (expectedReferences && expectedReferences.length > 0) {
+        await clickOnEmptyPartOfCodeView(driver)
+        await tokenEl.hover()
+        await waitForHover(driver, expectedHoverContains)
+        await (await driver.findElementWithText('Find references')).click()
+
+        await driver.page.waitForSelector('.e2e-search-result')
+        const refLinks = await collectLinks(driver)
+        for (const expectedReference of expectedReferences) {
+            expect(refLinks).toContainEqual(expectedReference)
+        }
+        await clickOnEmptyPartOfCodeView(driver)
+    }
+
+    // Go-to-definition
+    await clickOnEmptyPartOfCodeView(driver)
+    await tokenEl.hover()
+    await waitForHover(driver, expectedHoverContains)
+    await (await driver.findElementWithText('Go to definition')).click()
+
+    if (Array.isArray(expectedDefinition)) {
+        await driver.page.waitForSelector('.hierarchical-locations-view')
+        const defLinks = await collectLinks(driver)
+        for (const definition of expectedDefinition) {
+            expect(defLinks).toContainEqual(definition)
+        }
+    } else {
+        await driver.page.waitForFunction(
+            defURL => document.location.href.endsWith(defURL),
+            { timeout: 2000 },
+            expectedDefinition.url
+        )
+        await driver.page.goBack()
+    }
+
+    await driver.page.keyboard.press('Escape')
+}
+
+/**
+ * Return a list of locations (and their precision) that exist in the file list
+ * panel. This will click on each repository and collect the visible links in a
+ * sequence.
+ */
+async function collectLinks(driver: Driver): Promise<TestLocation[]> {
+    const panelTabTitles = await getPanelTabTitles(driver)
+    if (panelTabTitles.length === 0) {
+        return collectVisibleLinks(driver)
+    }
+
+    const links = []
+    for (const title of panelTabTitles) {
+        const tabElem = await driver.page.$$(`.e2e-hierarchical-locations-view-list span[title="${title}"]`)
+        if (tabElem.length > 0) {
+            await tabElem[0].click()
+        }
+
+        for (const link of await collectVisibleLinks(driver)) {
+            links.push(link)
+        }
+    }
+
+    return links
+}
+
+/**
+ * Return the list of repository titles on the left-hand side of the definition or
+ * reference result panel.
+ */
+async function getPanelTabTitles(driver: Driver): Promise<string[]> {
+    return (
+        await Promise.all(
+            (await driver.page.$$('.hierarchical-locations-view > div:nth-child(1) span[title]')).map(e =>
+                e.evaluate(e => e.getAttribute('title') || '')
+            )
+        )
+    ).map(normalizeWhitespace)
+}
+
+/**
+ * Return a list of locations (and their precision) that are current visible in a
+ * file list panel. This may be definitions or references.
+ */
+function collectVisibleLinks(driver: Driver): Promise<TestLocation[]> {
+    return driver.page.evaluate(() =>
+        Array.from(document.querySelectorAll<HTMLElement>('.e2e-search-result')).flatMap(e =>
+            Array.from(e.querySelectorAll<HTMLElement>('a[href]'))
+                .filter(a => a.nextElementSibling?.className.includes('file-match-children__item-badge-row'))
+                .map(a => ({
+                    url: a.getAttribute('href') || '',
+                    precise: a.nextElementSibling?.childElementCount === 0,
+                }))
+        )
+    )
+}
+
+/**
+ * Close any visible hover overlay.
+ */
+async function clickOnEmptyPartOfCodeView(driver: Driver): Promise<void> {
+    await driver.page.click('.e2e-blob tr:nth-child(1) .line')
+    await driver.page.waitForFunction(() => document.querySelectorAll('.e2e-tooltip-go-to-definition').length === 0)
+}
+
+/**
+ * Find the element with the token text on the given line.
+ *
+ * Will close any toast so that the enture line is visible and will hover over the line
+ * to ensure that the line is tokenized (as this is done on-demand).
+ */
+async function findTokenElement(driver: Driver, line: number, token: string): Promise<ElementHandle<Element>> {
+    try {
+        // If there's an open toast, close it. If the toast remains open and our target
+        // identifier happens to be hidden by it, we won't be able to select the correct
+        // token. This condition was reproducible in the code navigation test that searches
+        // for the identifier `StdioLogger`.
+        await driver.page.click('.e2e-close-toast')
+    } catch (error) {
+        // No toast open, this is fine
+    }
+
+    const selector = `.e2e-blob tr:nth-child(${line}) span`
+    await driver.page.hover(selector)
+    return driver.findElementWithText(token, { selector, fuzziness: 'exact' })
+}
+
+/**
+ * Wait for the hover tooltip to become visible. Compare the visible text with the expected
+ * contents (expected contents must be a substring of the visible contents). If precise is
+ * supplied, ensure that the presence of the UI indicator matches this value.
+ */
+async function waitForHover(driver: Driver, expectedHoverContains: string, precise?: boolean): Promise<void> {
+    await driver.page.waitForSelector('.e2e-tooltip-go-to-definition')
+    await driver.page.waitForSelector('.e2e-tooltip-content')
+    expect(normalizeWhitespace(await getTooltip(driver))).toContain(normalizeWhitespace(expectedHoverContains))
+
+    if (precise !== undefined) {
+        expect(
+            await driver.page.evaluate(() => document.querySelectorAll<HTMLElement>('.e2e-hover-badge').length)
+        ).toEqual(precise ? 0 : 1)
+    }
+}
+
+/**
+ * Return the currently visible hover text.
+ */
+async function getTooltip(driver: Driver): Promise<string> {
+    return driver.page.evaluate(() => (document.querySelector('.e2e-tooltip-content') as HTMLElement).innerText)
+}
+
+/**
+ * Collapse multiple spaces into one.
+ */
+function normalizeWhitespace(s: string): string {
+    return s.replace(/\s+/g, ' ')
+}
+
+//
+// LSIF utilities
+
+/**
+ * Replace the codeIntel.lsif setting with the given value.
+ *
+ * Return a promise that updates the global settings to their original value. This return value
+ * is suitable for use with the resource manager's destroy queue.
+ */
+export async function setGlobalLSIFSetting(gqlClient: GraphQLClient, enabled: boolean): Promise<() => Promise<void>> {
+    const { subjectID, settingsID, contents: oldContents } = await getGlobalSettings(gqlClient)
+    const newContents = applyEdits(
+        oldContents,
+        setProperty(oldContents, ['codeIntel.lsif'], enabled, {
+            eol: '\n',
+            insertSpaces: true,
+            tabSize: 2,
+        })
+    )
+
+    await overwriteSettings(gqlClient, subjectID, settingsID, newContents)
+    return async () => {
+        const { subjectID: currentSubjectID, settingsID: currentSettingsID } = await getGlobalSettings(gqlClient)
+        await overwriteSettings(gqlClient, currentSubjectID, currentSettingsID, oldContents)
+    }
+}
+
+/**
+ * Upload a single LSIF upload and ensure the upload is successfully processed
+ * and is visible from the tip of the default branch.
+ */
+export async function uploadAndEnsure(
+    driver: Driver,
+    config: Pick<Config, 'sourcegraphBaseUrl'>,
+    gqlClient: GraphQLClient,
+    repository: string,
+    commit: string,
+    root: string
+): Promise<() => Promise<void>> {
+    // First, remove all existing uploads for the repository
+    await clearUploads(gqlClient, repository)
+
+    // Upload each upload in parallel and get back the upload status URLs
+    const uploadUrl = await performUpload(config, {
+        repository,
+        commit,
+        root,
+        filename: `lsif-data/${repository}@${commit.substring(0, 12)}.lsif`,
+    })
+
+    // Check the upload status URLs to ensure that they succeed, then ensure
+    // that they are all listed as one of the "active" uploads for that repo
+    await ensureUpload(driver, config, {
+        repository,
+        commit,
+        root,
+        uploadUrl,
+    })
+
+    return (): Promise<void> => clearUploads(gqlClient, repository)
+}
+
+/**
+ * Delete all LSIF uploads for a repository.
+ */
+export async function clearUploads(gqlClient: GraphQLClient, repoName: string): Promise<void> {
+    const { nodes, hasNextPage } = await gqlClient
+        .queryGraphQL(
+            gql`
+                query ResolveRev($repoName: String!) {
+                    repository(name: $repoName) {
+                        lsifUploads {
+                            nodes {
+                                id
+                            }
+
+                            pageInfo {
+                                hasNextPage
+                            }
+                        }
+                    }
+                }
+            `,
+            { repoName }
+        )
+        .pipe(
+            map(dataOrThrowErrors),
+            map(({ repository }) =>
+                repository === null
+                    ? { nodes: [], hasNextPage: false }
+                    : { nodes: repository.lsifUploads.nodes, hasNextPage: repository.lsifUploads.pageInfo.hasNextPage }
+            )
+        )
+        .toPromise()
+
+    const indices = range(nodes.length)
+    const args: { [k: string]: string } = {}
+    indices.forEach(i => (args[`upload${i}`] = nodes[i].id))
+
+    await gqlClient
+        .mutateGraphQL(
+            gql`
+                mutation(${indices.map(i => `$upload${i}: ID!`).join(', ')}) {
+                    ${indices.map(i => gql`delete${i}: deleteLSIFUpload(id: $upload${i}) { alwaysNil }`).join('\n')}
+                }
+            `,
+            args
+        )
+        .pipe(map(dataOrThrowErrors))
+        .toPromise()
+
+    if (hasNextPage) {
+        // If we have more upload, clear the next page
+        return clearUploads(gqlClient, repoName)
+    }
+}
+
+/**
+ * Untar the LSIF data from the lsif-data directory and perform an LSIF upload
+ * via src command, which must be a binary available on the $PATH. Returns the
+ * URI of the upload's details (in UI).
+ */
+async function performUpload(
+    config: Pick<Config, 'sourcegraphBaseUrl'>,
+    {
+        repository,
+        commit,
+        root,
+        filename,
+    }: {
+        repository: string
+        commit: string
+        root: string
+        filename: string
+    }
+): Promise<string> {
+    const cwd = path.join(__dirname, 'lsif-data', path.dirname(repository))
+
+    try {
+        // Untar the lsif data for this upload
+        const tarCommand = ['tar', '-xzf', `${path.basename(filename)}.gz`].join(' ')
+        await child_process.exec(tarCommand, { cwd })
+    } catch (error) {
+        throw new Error(`Failed to untar test data: ${error}`)
+    }
+
+    let out!: string
+    try {
+        // Upload data
+        const uploadCommand = [
+            `src -endpoint ${config.sourcegraphBaseUrl}`,
+            'lsif upload',
+            `-repo ${repository}`,
+            `-commit ${commit}`,
+            `-root ${root}`,
+            `-file ${path.basename(filename)}`,
+        ].join(' ')
+        ;[out] = await child_process.exec(uploadCommand, { cwd })
+    } catch (error) {
+        try {
+            // See if the error is due to a missing utility
+            await child_process.exec('which src')
+        } catch (error) {
+            throw new Error('src-cli is not available on PATH')
+        }
+
+        throw new Error(`Failed to upload LSIF data: ${error.stderr || error.stdout || '(no output)'}`)
+    }
+
+    // Extract the status URL
+    const match = out.match(/To check the status, visit (.+).\n$/)
+    if (!match) {
+        throw new Error(`Unexpected output from Sourcegraph cli: ${out}`)
+    }
+
+    return match[1]
+}
+
+/**
+ * Refresh the upload page until it has finished processing. Then, navigate to the
+ * list of uploads for that repository and ensure that it's visible in the list of
+ * uploads visible at the tip of the default branch.
+ */
+async function ensureUpload(
+    driver: Driver,
+    config: Pick<Config, 'sourcegraphBaseUrl'>,
+    { repository, commit, root, uploadUrl }: { repository: string; commit: string; root: string; uploadUrl: string }
+): Promise<void> {
+    const pendingUploadStateMessages = ['Upload is queued.', 'Upload is currently being processed...']
+
+    await driver.page.goto(uploadUrl)
+    while (true) {
+        // Keep reloading upload page until the upload is terminal (not queued, not processed)
+        const text = await (await driver.page.waitForSelector('.e2e-upload-state')).evaluate(elem => elem.textContent)
+        if (!pendingUploadStateMessages.includes(text || '')) {
+            break
+        }
+
+        await driver.page.reload()
+    }
+
+    // Ensure upload is successful
+    const text = await (await driver.page.waitForSelector('.e2e-upload-state')).evaluate(elem => elem.textContent)
+    expect(text).toEqual('Upload completed successfully.')
+
+    await driver.page.goto(`${config.sourcegraphBaseUrl}/${repository}/-/settings/code-intelligence`)
+
+    const commitElem = await driver.page.waitForSelector('.e2e-upload-commit')
+    const actualCommit = await commitElem.evaluate(elem => elem.textContent)
+    expect(actualCommit).toEqual(commit.substr(0, 7))
+
+    const rootElem = await driver.page.waitForSelector('.e2e-upload-root')
+    const actualRoot = await rootElem.evaluate(elem => elem.textContent)
+    expect(actualRoot).toEqual(root)
+}
